Renaming encoder timestamp to mega timestamp to indicate to high level how long low level has been running
