+++ real_time/arduino_src/radio_buggy_mega/main.cpp
@@ -583,7 +583,7 @@ int main(void)
         g_rbsm.Send(RBSM_MID_MEGA_BATTERY_LEVEL, g_current_voltage);
         g_rbsm.Send(RBSM_MID_MEGA_STEER_FEEDBACK, steering_feedback_angle);
         g_rbsm.Send(RBSM_MID_ENC_TICKS_RESET, g_encoder_distance.GetTicks());
-        g_rbsm.Send(RBSM_MID_ENC_TIMESTAMP, millis());
+        g_rbsm.Send(RBSM_MID_MEGA_TIMESTAMP, millis());
         g_rbsm.Send(RBSM_MID_COMP_HASH, (long unsigned)(FP_HEXCOMMITHASH));
         g_rbsm.Send(RBSM_MID_ERROR, g_errors);
