+++ real_time/rbsm_config.txt
@@ -4,7 +4,6 @@
 
 RBSM_MID_ENC_TICKS_LAST, 0
 RBSM_MID_ENC_TICKS_RESET, 1
-RBSM_MID_ENC_TIMESTAMP, 2
 RBSM_MID_ENC_RESET_REQUEST, 5
 RBSM_MID_ENC_RESET_CONFIRM, 6
 RBSM_MID_MEGA_TELEOP_BRAKE_COMMAND, 17
@@ -16,6 +15,7 @@ RBSM_MID_MEGA_AUTON_STATE, 22
 RBSM_MID_MEGA_BATTERY_LEVEL, 23
 RBSM_MID_MEGA_STEER_FEEDBACK, 24
 RBSM_MID_COMP_HASH, 30
+RBSM_MID_MEGA_TIMESTAMP, 253
 RBSM_MID_ERROR, 254
 DEVICE_ID, 255
 FOOTER, 10
