+++ real_time/surface_src/java_src/Alice/src/main/java/com/roboclub/robobuggy/messages/EncoderResetMessage.java
@@ -17,7 +17,7 @@ public class EncoderResetMessage extends BaseMessage {
 	public static final String VERSION_ID = "EncoderResetV0.1";
 
 	/**
-	 * Construct a new {@link EncoderTimeMessage} at time now
+	 * Construct a new {@link MegaTimeMessage} at time now
 	 */
 	public EncoderResetMessage() {
 		this.timestamp = new Date().getTime();
